requirements.txt
Django==2.0.1
django-wkhtmltopdf==3.2.0
